refactor: extract breadth backfill executor
